refactor: Move API functions into hooks and de-couple
