+++ frontend/src/components/sign-in.tsx
@@ -1,34 +1,12 @@
 // libs
 import { Button, Dialog, TextField } from "@radix-ui/themes";
-import { useQuery } from "@tanstack/react-query";
 import { ChangeEvent, useState } from "react";
 import { useShallow } from "zustand/react/shallow";
 
 // internals
 import { isSurvivor } from "../helpers";
-import { Toast, useSurvivorStore, useToastsStore } from "../stores";
-import { ErrorState, Survivor } from "../types";
-
-const TOASTS: Record<string, Toast> = {
-  "signin-not-fount": {
-    title: "Error",
-    description: "Could not find anyone in the system with that name",
-    type: "error",
-    open: false,
-  },
-  "signin-error": {
-    title: "Error",
-    description: "An error occurred while signing in",
-    type: "error",
-    open: false,
-  },
-  "signin-data-corrupted": {
-    title: "Error",
-    description: "Your profile data is corrupted",
-    type: "error",
-    open: false,
-  },
-};
+import { useSurvivorStore } from "../stores";
+import { useSurvivor } from "../hooks";
 
 export function SignIn() {
   const { identify, updateInventory } = useSurvivorStore(
@@ -38,38 +16,28 @@ export function SignIn() {
     })),
   );
 
-  const { openToast, bulkRegisterToasts } = useToastsStore(
-    useShallow((state) => ({
-      openToast: state.actions.openToast,
-      bulkRegisterToasts: state.actions.bulkRegisterToasts,
-    })),
-  );
-  // Avoid updating ToastsEngine while this renders
-  setTimeout(() => bulkRegisterToasts({ ...TOASTS }), 0);
-
   const [name, setName] = useState("");
 
-  const { refetch, isFetching } = useQuery<Survivor | ErrorState>({
-    queryKey: ["get-survivor", name],
-    queryFn: async () => {
-      const response = await fetch(`/api/survivors/${name}`);
-
-      if (!response.ok) {
-        if (response.status === 404) {
-          openToast("signin-not-fount");
-          throw new Error("Not found");
-        }
-        openToast("signin-error");
-        throw new Error("An error occurred");
-      }
-      try {
-        return await response.json();
-      } catch (err) {
-        openToast("signin-data-corrupted");
-        throw err;
-      }
-    },
+  const { isFetching, refetch } = useSurvivor({
+    identifier: name,
     enabled: false,
+    onSuccess: (data) => {
+      if (
+        isSurvivor(data) &&
+        data.name.toLocaleLowerCase() === name.toLocaleLowerCase() &&
+        data.id
+      ) {
+        identify(
+          data.id,
+          data.name,
+          data.lastLocation.latitude,
+          data.lastLocation.longitude,
+        );
+        updateInventory(data.inventory);
+      }
+
+      setName("");
+    },
   });
 
   return (
@@ -103,24 +71,7 @@ export function SignIn() {
         </Dialog.Close>
         <Dialog.Close>
           <Button
-            onClick={async () => {
-              const { data } = await refetch();
-              if (
-                isSurvivor(data) &&
-                data.name.toLocaleLowerCase() === name.toLocaleLowerCase() &&
-                data.id
-              ) {
-                identify(
-                  data.id,
-                  data.name,
-                  data.lastLocation.latitude,
-                  data.lastLocation.longitude,
-                );
-                updateInventory(data.inventory);
-              }
-
-              setName("");
-            }}
+            onClick={() => refetch()}
             disabled={name === ""}
             loading={name !== "" && isFetching}
           >
