+++ frontend/src/components/trading-panel.tsx
@@ -1,34 +1,15 @@
 // libs
 import { Button, Dialog } from "@radix-ui/themes";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { useCallback, useMemo, useState } from "react";
+import { useQueryClient } from "@tanstack/react-query";
+import { useMemo, useState } from "react";
 import { useShallow } from "zustand/react/shallow";
 
 // internals
 import { Inventory } from ".";
-import { useItems } from "../hooks";
-import {
-  Toast,
-  useSearchStore,
-  useSurvivorStore,
-  useToastsStore,
-} from "../stores";
+import { QueryKeys } from "../constants";
+import { useItems, useTrade } from "../hooks";
+import { useSearchStore, useSurvivorStore, useToastsStore } from "../stores";
 import { Survivor, Inventory as TInventory } from "../types";
-
-const TOASTS: Record<string, Toast> = {
-  "trade-error": {
-    title: "Trade failed",
-    description: "An error occurred while trading items",
-    type: "error",
-    open: false,
-  },
-  "trade-success": {
-    title: "Trade successful",
-    description: "Items have been traded successfully",
-    type: "success",
-    open: false,
-  },
-};
 
 interface TradingPanelProps {
   tradingPartner: Survivor;
@@ -40,17 +21,30 @@ export function TradingPanel({ tradingPartner }: TradingPanelProps) {
       myInventory: state.inventory,
     })),
   );
-
+  const openToast = useToastsStore((state) => state.actions.openToast);
   const maxDistance = useSearchStore((state) => state.maxDistance);
 
-  const { openToast, bulkRegisterToasts } = useToastsStore(
-    useShallow((state) => ({
-      openToast: state.actions.openToast,
-      bulkRegisterToasts: state.actions.bulkRegisterToasts,
-    })),
-  );
-  // Avoid updating ToastsEngine while this renders
-  setTimeout(() => bulkRegisterToasts({ ...TOASTS }), 0);
+  const { mutate } = useTrade({
+    identifier: myId,
+    onSuccess: async () => {
+      openToast("trade-success");
+
+      setMyTempInventory({
+        ...myInventory,
+      });
+      setMyOffer(undefined);
+      setTheirInventory({
+        ...tradingPartner.inventory,
+      });
+      setTheirOffer(undefined);
+      queryClient.invalidateQueries({
+        queryKey: [QueryKeys.GetSurvivor, myId],
+      });
+      queryClient.invalidateQueries({
+        queryKey: [QueryKeys.GetSurvivors, myId, maxDistance],
+      });
+    },
+  });
 
   const possibleItems = useItems();
 
@@ -82,62 +76,6 @@ export function TradingPanel({ tradingPartner }: TradingPanelProps) {
   };
 
   const queryClient = useQueryClient();
-
-  const mutationFn = useCallback(
-    async (data: {
-      survivor_a_items: { survivor_id: string; items: TInventory };
-      survivor_b_items: { survivor_id: string; items: TInventory };
-    }) => {
-      const headers = new Headers();
-      headers.append("Content-Type", "application/json");
-      headers.append("X-User-Id", myId!);
-
-      const response = await fetch("/api/survivors/trade/", {
-        method: "POST",
-        headers,
-        body: JSON.stringify(data),
-      });
-
-      if (!response.ok) {
-        openToast("trade-error");
-        throw new Error("Failed to trade");
-      }
-      try {
-        return await response.json();
-      } catch (err) {
-        openToast("trade-error");
-        throw err;
-      }
-    },
-    [myId, openToast],
-  );
-
-  const mutation = useMutation({
-    mutationFn,
-    onSuccess: async () => {
-      openToast("trade-success");
-
-      setMyTempInventory({
-        ...myInventory,
-      });
-      setMyOffer(undefined);
-      setTheirInventory({
-        ...tradingPartner.inventory,
-      });
-      setTheirOffer(undefined);
-
-      Promise.all([
-        // Refresh my profile
-        queryClient.invalidateQueries({
-          queryKey: ["get-survivor", myId],
-        }),
-        // Refresh full list
-        queryClient.invalidateQueries({
-          queryKey: ["get-survivors", myId, maxDistance],
-        }),
-      ]);
-    },
-  });
 
   return (
     <Dialog.Content>
@@ -201,15 +139,15 @@ export function TradingPanel({ tradingPartner }: TradingPanelProps) {
         <Dialog.Close>
           <Button
             disabled={!tradingPartner.id || !myOffer || !theirOffer}
-            onClick={async () => {
-              await mutation.mutate({
+            onClick={() =>
+              mutate({
                 survivor_a_items: { survivor_id: myId!, items: myOffer! },
                 survivor_b_items: {
                   survivor_id: tradingPartner.id!,
                   items: theirOffer!,
                 }!,
-              });
-            }}
+              })
+            }
           >
             Trade
           </Button>
