+++ frontend/src/components/user-control.tsx
@@ -6,38 +6,16 @@ import {
   Text,
   VisuallyHidden,
 } from "@radix-ui/themes";
-import { useQuery } from "@tanstack/react-query";
 import { useShallow } from "zustand/react/shallow";
 
 // internals
 import { Register, SignIn, UserPanel } from ".";
 import { isSurvivor } from "../helpers";
-import { Toast, useSurvivorStore, useToastsStore } from "../stores";
-import { ErrorState, Survivor } from "../types";
-
-const TOASTS: Record<string, Toast> = {
-  "loadprofile-not-fount": {
-    title: "Error",
-    description: "Could not find your profile in the system",
-    type: "error",
-    open: false,
-  },
-  "loadprofile-error": {
-    title: "Error",
-    description: "An error occurred while loading your profile",
-    type: "error",
-    open: false,
-  },
-  "loadprofile-data-corrupted": {
-    title: "Error",
-    description: "Your profile data is corrupted",
-    type: "error",
-    open: false,
-  },
-};
+import { useSurvivor } from "../hooks";
+import { useSurvivorStore } from "../stores";
 
 export function UserCenter() {
-  const { myId, myName, identify, updateInventory } = useSurvivorStore(
+  const { identify, myId, myName, updateInventory } = useSurvivorStore(
     useShallow((state) => ({
       myId: state.id,
       myName: state.name,
@@ -46,47 +24,20 @@ export function UserCenter() {
     })),
   );
 
-  const { openToast, bulkRegisterToasts } = useToastsStore(
-    useShallow((state) => ({
-      openToast: state.actions.openToast,
-      bulkRegisterToasts: state.actions.bulkRegisterToasts,
-    })),
-  );
-  // Avoid updating ToastsEngine while this renders
-  setTimeout(() => bulkRegisterToasts({ ...TOASTS }), 0);
-
-  const { isFetching } = useQuery<Survivor | ErrorState>({
-    queryKey: ["get-survivor", myId],
-    queryFn: async () => {
-      const response = await fetch(`/api/survivors/${myId}`);
-
-      if (!response.ok) {
-        if (response.status === 404) {
-          openToast("loadprofile-not-fount");
-          throw new Error("Not found");
-        }
-        openToast("loadprofile-error");
-        throw new Error("An error occurred");
-      }
-
-      try {
-        const data = await response.json();
-        if (isSurvivor(data) && data.id === myId) {
-          identify(
-            data.id,
-            data.name,
-            data.lastLocation.latitude,
-            data.lastLocation.longitude,
-          );
-          updateInventory(data.inventory);
-        }
-        return data;
-      } catch (err) {
-        openToast("loadprofile-data-corrupted");
-        throw err;
+  const { isFetching } = useSurvivor({
+    identifier: myId,
+    enabled: !!myId,
+    onSuccess: (data) => {
+      if (isSurvivor(data) && data.id === myId) {
+        identify(
+          data.id,
+          data.name,
+          data.lastLocation.latitude,
+          data.lastLocation.longitude,
+        );
+        updateInventory(data.inventory);
       }
     },
-    enabled: !!myId,
   });
 
   if (myId && !myName && isFetching) {
