+++ frontend/src/hooks/index.ts
@@ -1,1 +1,7 @@
+export * from "./use-create-survivor.hook";
 export * from "./use-items.hook";
+export * from "./use-report-infection.hook";
+export * from "./use-survivor.hook";
+export * from "./use-survivors.hook";
+export * from "./use-trade.hook";
+export * from "./use-update-survivor.hook";
